now with more goats

+++ cicd-tests/remote-senspot-file-test.sh
@@ -11,14 +11,15 @@ echo "zzzfile created"
 ssh ubuntu@$ADDR "cd /home/ubuntu/cspot/build/bin && ./senspot-file-init -W zzzfilebig -M 64 -s 10"
 echo "zzzfilebig created"
 # create test file
-rm -f ./zzzfff ./zzzxxx
+rm -f ./zzzfff ./zzzxxx ./zzzaaa
 cp ../../cicd-tests/zzzfff .
 cp ../../cicd-tests/zzzxxx .
+cp ../../cicd-tests/zzzaaa .
 
 FTEST=`./senspot-file-send -f ./zzzfff -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfile -V | grep "megabytes"`
 if ( test -z "$FTEST" ) ; then
 	echo "senspot-file-send failed"
-	rm -f ./zzzfff ./zzzxxx
+	rm -f ./zzzfff ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -28,7 +29,7 @@ RTEST=`./senspot-file-recv -f ./zzzggg -W woof://$ADDR/home/ubuntu/cspot/build/b
 
 if ( test -z "$RTEST" ) ; then
 	echo "senspot-file-recv failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -38,7 +39,7 @@ DTEST=`diff ./zzzfff ./zzzggg`
 
 if ( ! test -z "$DTEST" ) ; then
 	echo "file diff failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -47,7 +48,7 @@ fi
 FTEST=`./senspot-file-send -f ./zzzfff -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfilebig -V | grep "megabytes"`
 if ( test -z "$FTEST" ) ; then
 	echo "senspot-file-send big failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -57,7 +58,7 @@ RTEST=`./senspot-file-recv -f ./zzzggg -W woof://$ADDR/home/ubuntu/cspot/build/b
 
 if ( test -z "$RTEST" ) ; then
 	echo "senspot-file-recv big failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -67,7 +68,7 @@ DTEST=`diff ./zzzfff ./zzzggg`
 
 if ( ! test -z "$DTEST" ) ; then
 	echo "file diff big failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -75,14 +76,14 @@ fi
 
 # try multi-write and versioning
 ./senspot-file-send -f ./zzzfff -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfile &
-./senspot-file-send -f ./zzzfff -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfile &
+./senspot-file-send -f ./zzzaaa -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfile &
 
 sleep 1
 FTEST=`./senspot-file-send -f ./zzzxxx -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfilebig -V | grep "megabytes"`
 
 if ( test -z "$FTEST" ) ; then
 	echo "second file send failed"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -97,7 +98,7 @@ echo "fetching version $VER"
 RTEST=`./senspot-file-recv -f ./zzzyyy -W woof://$ADDR/home/ubuntu/cspot/build/bin/zzzfile -v $MAJ -m $MIN -V | grep "megabytes"`
 if ( test -z "$RTEST" ) ; then
 	echo "coupld not fetch version $VER"
-	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy
+	rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy ./zzzaaa
 	ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
 	$(pwd)/kill-remote-platform.sh 
 	exit 1
@@ -106,19 +107,22 @@ fi
 
 DTEST1=`diff ./zzzfff ./zzzyyy`
 DTEST2=`diff ./zzzxxx ./zzzyyy`
+DTEST3=`diff ./zzzaaa ./zzzyyy`
 if ( ! test -z "$DTEST1" ) ; then
 	if ( ! test -z "$DTEST2" ) ; then
-		echo "diff fails for multi version"
-		rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy
-		ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
-		$(pwd)/kill-remote-platform.sh 
-		exit 1
+		if ( ! test -z "$DTEST3" ) ; then
+			echo "diff fails for multi version"
+			rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy ./zzzaaa
+			ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
+			$(pwd)/kill-remote-platform.sh 
+			exit 1
+		fi
 	fi
 fi
 
 $(pwd)/kill-remote-platform.sh 
 ssh ubuntu@$ADDR "rm -f /home/ubuntu/cspot/build/bin/zzzfile*"
-rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy
+rm -f ./zzzfff ./zzzggg ./zzzxxx ./zzzyyy ./zzzaaa
 #echo "sending HUP to $WPID"
 #
 #
